Update dtype and enable fused AdamW
- Change `dtype` from `bfloat16` to `float32` in `train_daggpt.py` for better precision.
- Uncomment and enable use of fused AdamW optimizer in `dag_model.py` to improve performance.
- Add print statement to log whether fused AdamW is being used.

+++ config/train_daggpt.py
@@ -35,7 +35,8 @@ min_lr = 3e-5
 dataset = "proofpile"
 
 backend = "nccl"
-dtype = "bfloat16"
+# dtype = "bfloat16"
+dtype = "float32"
 compile = False  # Currently broken for dag model
 
 always_save_checkpoint = True
